Update main.c - Clear Bottom Screen
Clear bottom screen to black on boot. It's an eye sore. Especially with the new GUI. Tested this and it works! This will clear bottom screen and set it to black. If you ever decide to put a logo there, you can remove this. Otherwise this should do well in the mean time!