Forked and changed re-added.
Re-added changes to fork of mid-kid's fork of rxTools. Now uses CakeHax
and does away with gatway's rop.

Some tweaks to top screen splash logo as well. (text on bottom now
handled by the menu system)

Note this versions uses firmwareRX.bin instead of firmware.bin so as to
not conflict with CakesFW.